feat(explode): Extract Sub Class
- Extract all sub classes to same package
- create a Renderer class
- create a Config class with argument analyzer
- Fix issues on MapReader and MapRenderer
