+++ src/main/java/GameObjectType.java
@@ -1,0 +1,6 @@
+/**
+ * Game object type definition.
+ */
+public enum GameObjectType {
+    RECTANGLE, CIRCLE, IMAGE
+}
